Update config.plist
372 - 332 has been edited;

AAPL,ig-platform-id: The value was changed to "07009B3E" to use the correct platform ID for Intel UHD Graphics 630 on Comet Lake. The previous value "BwCbPg==" corresponded to the platform ID for Skylake.

device-id: The value was changed to "9BC80000" to correspond with the correct device ID for Intel UHD Graphics 630 on Comet Lake. The previous value "mz4AAA==" was not the correct ID.

framebuffer-con0-enable: The value was changed to "AQAAAA==" to enable the first video output of the integrated graphics controller. The previous value "AgAAAA==" disabled the output.

framebuffer-con0-pipe: The value was changed to "AAAAAAAA" to select the correct pipe for the first video output of the integrated graphics controller. The previous value "EgAAAA==" was not the correct pipe.

framebuffer-con0-type: The value was changed to "AAAAAA==" to select the correct connection type for the first video output of the integrated graphics controller. The previous value "AAgAAA==" was not the correct type.

framebuffer-fbmem: The value was changed to "00000080" to allocate 128 MB of video memory to the integrated graphics controller. The previous value "AACQAA==" allocated 160 MB, which can cause instability issues.

framebuffer-patch-enable: The value was changed to "AQAAAA==" to enable the framebuffer patch for the integrated graphics controller. The previous value "AgAAAA==" disabled the patch.

framebuffer-stolenmem: The value was changed to "00003001" to allocate 48 MB of stolen memory to the integrated graphics controller. The previous value "AAAwAQ==" allocated 9 MB, which can cause instability issues.

framebuffer-unifiedmem: The value was changed to "00000080" to allocate 128 MB of unified memory to the integrated graphics controller. The previous value "AAAAgA==" allocated 256 MB, which can cause instability issues.

hda-gfx: The value was changed to "onboard-1" to assign the HDMI audio output to the integrated graphics controller. The previous value "onboard-2" was incorrect.

model: The value was changed to "Intel UHD Graphics 630" to correctly identify the model of the integrated graphics controller. The previous value "Intel CoffeeLake-H GT2 [UHD Graphics 630]" had additional information that was unnecessary and could cause issues.

Tested on a Dell Vostro 3681 that I bought in Arizona back on 2020, I can't tell if there is a variant with the CN version, if that is the case, the following commands are used to obtain information about the computer hardware, specifically for the integrated graphics card, and the VGA and HDMI ports.

The first command, "ioreg -l | grep 'device-id'", displays the device ID of the integrated graphics card. This can be useful for identifying the graphics card model and finding specific drivers.

The second command, "ioreg -lw0 | grep 'IODisplayEDID' | sed '/[^<]*</s///' | xxd -p -r | egrep -c 'CMO|Chimei|Philips|SAMSUNG'", shows the number of displays connected to the VGA port and their manufacturer. It uses the information stored in the Extended Display Identification Data (EDID) to identify the manufacturer of the display.

The third command, "ioreg -lw0 | grep 'IODisplayEDID' | sed '/[^<]*</s///' | xxd -p -r | egrep -c 'CEC|DVI|HDMI'", is similar to the second command, but for the HDMI port instead of the VGA port. It also uses the EDID to identify the type of connection of the display.

In summary, these commands can provide useful information about the computer hardware, which can be helpful for troubleshooting compatibility issues or finding specific drivers for the hardware.